requirements.txt
numpy
opencv-python
matplotlib
pandas
pytest
torch
transformers
accelerate
Pillow
ultralytics
mediapipe
